Add GraphQL tests for Export operations
